Fixed non-responsive tools in IE/FF, issue #6
FF and IE had issues with non-responsive clicks for tools.
The View was not updated automatically upon firing of the functions,
thus the user had to fire the function(e.g rotate) and then move his
mouse 1px in any direction for the function to take effect on the
canvas. This commit fixes this by calling the view.update() manually
within each function.

https://github.com/nicholaswmin/Vesper.next/issues/6