Added WheelZoom, ZoomRect, Default Opacity
I added a mouse wheel zoom which is activated when the user scrolls
in/out. Doesn't need firing up since it's always on, even when drawing
paths.

Modified zoom-rectangle to be activated through Space+CTRL

Added default opacity to Pen Tool that takes it's value from the
globalVariables.js file.